WFCORE-3594 Reload is racey with regard to reading the process state
Presently when issuing a reload operation the process state is updated asyncrounously,
which means that it is possible for a subsequent read of the process-state to return
a state of 'running' even though the server is actually in the process of reloading.

This can cause intermittent failures in any tests that perform a reload, as the general
pattern followed by these tests is to do a reload and then check the process state in a
loop until it is back to 'running'.